packages/core: rename api dir to api-wrappers
